Update madgraph to MG5_aMC_v2_3_3

+++ madgraph/madgraph/various/misc.py
@@ -15,6 +15,7 @@
 
 """A set of functions performing routine administrative I/O tasks."""
 
+import contextlib
 import logging
 import os
 import re
@@ -26,20 +27,24 @@ import sys
 import optparse
 import time
 import shutil
+import traceback
 import gzip as ziplib
 
 try:
     # Use in MadGraph
     import madgraph
-    from madgraph import MadGraph5Error, InvalidCmd
-    import madgraph.iolibs.files as files
 except Exception, error:
-    if __debug__:
-        print error
     # Use in MadEvent
-    import internal as madgraph
+    import internal
     from internal import MadGraph5Error, InvalidCmd
     import internal.files as files
+    MADEVENT = True    
+else:
+    from madgraph import MadGraph5Error, InvalidCmd
+    import madgraph.iolibs.files as files
+    MADEVENT = False
+
+
     
 logger = logging.getLogger('cmdprint.ext_program')
 logger_stderr = logging.getLogger('madevent.misc')
@@ -106,14 +111,19 @@ def get_pkg_info(info_str=None):
     string can be passed to be read instead of the file content.
     """
 
-    if info_str is None:
+    if info_str:
+        info_dict = parse_info_str(StringIO.StringIO(info_str))
+
+    elif MADEVENT:
+        info_dict ={}
+        info_dict['version'] = open(pjoin(internal.__path__[0],'..','..','MGMEVersion.txt')).read().strip()
+        info_dict['date'] = '20xx-xx-xx'                        
+    else:
         info_dict = files.read_from_file(os.path.join(madgraph.__path__[0],
                                                   "VERSION"),
                                                   parse_info_str, 
                                                   print_error=False)
-    else:
-        info_dict = parse_info_str(StringIO.StringIO(info_str))
-
+        
     return info_dict
 
 #===============================================================================
@@ -168,6 +178,68 @@ def which(program):
                 return exe_file
     return None
 
+def has_f2py():
+    has_f2py = False
+    if which('f2py'):
+        has_f2py = True
+    elif sys.version_info[1] == 6:
+        if which('f2py-2.6'):
+            has_f2py = True
+        elif which('f2py2.6'):
+            has_f2py = True                 
+    else:
+        if which('f2py-2.7'):
+            has_f2py = True 
+        elif which('f2py2.7'):
+            has_f2py = True  
+    return has_f2py       
+        
+#===============================================================================
+#  Activate dependencies if possible. Mainly for tests
+#===============================================================================
+
+def deactivate_dependence(dependency, cmd=None, log = None):
+    """ Make sure to turn off some dependency of MG5aMC. """
+    
+    def tell(msg):
+        if log == 'stdout':
+            print msg
+        elif callable(log):
+            log(msg)
+    
+
+    if dependency in ['pjfry','golem']:
+        if cmd.options[dependency] not in ['None',None,'']:
+            tell("Deactivating MG5_aMC dependency '%s'"%dependency)
+            cmd.options[dependency] = 'None'
+
+def activate_dependence(dependency, cmd=None, log = None):
+    """ Checks whether the specfieid MG dependency can be activated if it was
+    not turned off in MG5 options."""
+    
+    def tell(msg):
+        if log == 'stdout':
+            print msg
+        elif callable(log):
+            log(msg)
+
+    if cmd is None:
+        cmd = MGCmd.MasterCmd()
+
+    if dependency=='pjfry':
+        if cmd.options['pjfry'] in ['None',None,''] or \
+         (cmd.options['pjfry'] == 'auto' and which_lib('libpjfry.a') is None) or\
+          which_lib(pjoin(cmd.options['pjfry'],'libpjfry.a')) is None:
+            tell("Installing PJFry...")
+            cmd.do_install('PJFry')
+
+    if dependency=='golem':
+        if cmd.options['golem'] in ['None',None,''] or\
+         (cmd.options['golem'] == 'auto' and which_lib('libgolem.a') is None) or\
+         which_lib(pjoin(cmd.options['golem'],'libgolem.a')) is None:
+            tell("Installing Golem95...")
+            cmd.do_install('Golem95')
+    
 #===============================================================================
 # find a library
 #===============================================================================
@@ -236,12 +308,42 @@ def multiple_try(nb_try=5, sleep=20):
                         logger_stderr.debug('fail to do %s function with %s args. %s try on a max of %s (%s waiting time)' %
                                  (str(f), ', '.join([str(a) for a in args]), i+1, nb_try, sleep * (i+1)))
                         logger_stderr.debug('error is %s' % str(error))
+                        if __debug__: logger_stderr.debug('and occurred at :'+traceback.format_exc())
                     wait_once = True
                     time.sleep(sleep * (i+1))
 
+            if __debug__:
+                raise
             raise error.__class__, '[Fail %i times] \n %s ' % (i+1, error)
         return deco_f_retry
     return deco_retry
+
+#===============================================================================
+# helper for scan. providing a nice formatted string for the scan name
+#===============================================================================
+def get_scan_name(first, last):
+    """return a name of the type xxxx[A-B]yyy
+        where xxx and yyy are the common part between the two names.
+    """
+    
+    # find the common string at the beginning     
+    base = [first[i] for i in range(len(first)) if first[:i+1] == last[:i+1]]
+    # remove digit even if in common
+    while base[-1].isdigit():
+        base = base[:-1] 
+    # find the common string at the end 
+    end = [first[-(i+1)] for i in range(len(first)) if first[-(i+1):] == last[-(i+1):]]
+    # remove digit even if in common    
+    while end and end[-1].isdigit():
+        end = end[:-1] 
+    end.reverse()
+    #convert to string
+    base, end = ''.join(base), ''.join(end)
+    if end:
+        name = "%s[%s-%s]%s" % (base, first[len(base):-len(end)], last[len(base):-len(end)],end)
+    else:
+        name = "%s[%s-%s]%s" % (base, first[len(base):], last[len(base):],end)
+    return name
 
 #===============================================================================
 # Compiler which returns smart output error in case of trouble
@@ -351,6 +453,8 @@ def mod_compilator(directory, new='gfortran', current=None, compiler_type='gfort
                 lines[iline] = result.group(1) + var + "=" + new
         if mod:
             open(name,'w').write('\n'.join(lines))
+            # reset it to change the next file
+            mod = False
 
 #===============================================================================
 # mute_logger (designed to work as with statement)
@@ -441,6 +545,49 @@ class MuteLogger(object):
             #for i, h in enumerate(my_logger.handlers):
             #    h.setLevel(cls.logger_saved_info[logname][2][i])
 
+nb_open =0
+@contextlib.contextmanager
+def stdchannel_redirected(stdchannel, dest_filename):
+    """                                                                                                                                                                                                     
+    A context manager to temporarily redirect stdout or stderr                                                                                                                                              
+                                                                                                                                                                                                            
+    e.g.:                                                                                                                                                                                                   
+                                                                                                                                                                                                            
+                                                                                                                                                                                                            
+    with stdchannel_redirected(sys.stderr, os.devnull):                                                                                                                                                     
+        if compiler.has_function('clock_gettime', libraries=['rt']):                                                                                                                                        
+            libraries.append('rt')                                                                                                                                                                          
+    """
+
+    try:
+        oldstdchannel = os.dup(stdchannel.fileno())
+        dest_file = open(dest_filename, 'w')
+        os.dup2(dest_file.fileno(), stdchannel.fileno())
+        yield
+    finally:
+        if oldstdchannel is not None:
+            os.dup2(oldstdchannel, stdchannel.fileno())
+            os.close(oldstdchannel)
+        if dest_file is not None:
+            dest_file.close()
+        
+def get_open_fds():
+    '''
+    return the number of open file descriptors for current process
+
+    .. warning: will only work on UNIX-like os-es.
+    '''
+    import subprocess
+    import os
+
+    pid = os.getpid()
+    procs = subprocess.check_output( 
+        [ "lsof", '-w', '-Ff', "-p", str( pid ) ] )
+    nprocs = filter( 
+            lambda s: s and s[ 0 ] == 'f' and s[1: ].isdigit(),
+            procs.split( '\n' ) )
+        
+    return nprocs
 
 def detect_current_compiler(path, compiler_type='fortran'):
     """find the current compiler for the current directory"""
@@ -493,6 +640,19 @@ def rm_old_compile_file():
                                  if os.path.exists(os.path.join(lib_pos, lib))]
 
 
+def format_time(n_secs):
+    m, s = divmod(n_secs, 60)
+    h, m = divmod(m, 60)
+    d, h = divmod(h, 24)
+    if d > 0:
+        return "%d day%s,%dh%02dm%02ds" % (d,'' if d<=1 else 's',h, m, s)
+    elif h > 0:
+        return "%dh%02dm%02ds" % (h, m, s)
+    elif m > 0:
+        return "%dm%02ds" % (m, s)                
+    else:
+        return "%d second%s" % (s, '' if s<=1 else 's')   
+
 def rm_file_extension( ext, dirname, names):
 
     [os.remove(os.path.join(dirname, name)) for name in names if name.endswith(ext)]
@@ -543,8 +703,12 @@ def check_system_error(value=1):
 @check_system_error()
 def call(arg, *args, **opt):
     """nice way to call an external program with nice error treatment"""
-    return subprocess.call(arg, *args, **opt)
-
+    try:
+        return subprocess.call(arg, *args, **opt)
+    except OSError:
+        arg[0] = './%s' % arg[0]
+        return subprocess.call(arg, *args, **opt)
+        
 @check_system_error()
 def Popen(arg, *args, **opt):
     """nice way to call an external program with nice error treatment"""
@@ -686,6 +850,10 @@ class TMP_directory(object):
 
     
     def __exit__(self, ctype, value, traceback ):
+        #True only for debugging:
+        if False and isinstance(value, Exception):
+            sprint("Directory %s not cleaned. This directory can be removed manually" % self.path)
+            return False
         try:
             shutil.rmtree(self.path)
         except OSError:
@@ -698,6 +866,25 @@ class TMP_directory(object):
         
     def __enter__(self):
         return self.path
+    
+class TMP_variable(object):
+    """create a temporary directory and ensure this one to be cleaned.
+    """
+
+    def __init__(self, cls, attribute, value):
+        
+        self.old_value = getattr(cls, attribute)
+        self.cls = cls
+        self.attribute = attribute
+        setattr(self.cls, self.attribute, value)
+    
+    def __exit__(self, ctype, value, traceback ):
+
+        setattr(self.cls, self.attribute, self.old_value)
+        
+    def __enter__(self):
+        return self.old_value 
+    
 #
 # GUNZIP/GZIP
 #
@@ -721,8 +908,21 @@ def gunzip(path, keep=False, stdout=None):
         return 0
     
     if not stdout:
-        stdout = path[:-3]        
-    open(stdout,'w').write(ziplib.open(path, "r").read())
+        stdout = path[:-3]
+    try:
+        gfile = ziplib.open(path, "r")
+    except IOError:
+        raise
+    else:    
+        try:    
+            open(stdout,'w').write(gfile.read())
+        except IOError:
+            # this means that the file is actually not gzip
+            if stdout == path:
+                return
+            else:
+                files.cp(path, stdout)
+            
     if not keep:
         os.remove(path)
     return 0
@@ -919,6 +1119,10 @@ class OptionParser(optparse.OptionParser):
 
 def sprint(*args, **opt):
     """Returns the current line number in our program."""
+    
+    if not __debug__:
+        return
+    
     import inspect
     if opt.has_key('log'):
         log = opt['log']
@@ -928,18 +1132,36 @@ def sprint(*args, **opt):
         level = opt['level']
     else:
         level = logging.getLogger('madgraph').level
-        if level == 20:
-            level = 10 #avoid info level
+        #print  "madgraph level",level
+        #if level == 20:
+        #    level = 10 #avoid info level
+        #print "use", level
     lineno  =  inspect.currentframe().f_back.f_lineno
     fargs =  inspect.getframeinfo(inspect.currentframe().f_back)
     filename, lineno = fargs[:2]
     #file = inspect.currentframe().f_back.co_filename
     #print type(file)
+    try:
+        source = inspect.getsourcelines(inspect.currentframe().f_back)
+        line = source[0][lineno-source[1]]
+        line = re.findall(r"misc\.sprint\(\s*(.*)\)\s*($|#)", line)[0][0]
+        if line.startswith("'") and line.endswith("'") and line.count(",") ==0:
+            line= ''
+        elif line.startswith("\"") and line.endswith("\"") and line.count(",") ==0:
+            line= ''
+        elif line.startswith(("\"","'")) and len(args)==1 and "%" in line:
+            line= ''        
+    except Exception:
+        line=''
 
-
-    log.log(level, ' '.join([str(a) for a in args]) + \
-               '\nraised at %s at line %s ' % (filename, lineno))
+    if line:
+        intro = ' %s = \033[0m' % line
+    else:
+        intro = ''
     
+
+    log.log(level, ' '.join([intro]+[str(a) for a in args]) + \
+                   ' \033[1;30m[%s at line %s]\033[0m' % (os.path.basename(filename), lineno))
     return 
 
 ################################################################################
@@ -973,6 +1195,29 @@ class chdir:
     def __exit__(self, etype, value, traceback):
         os.chdir(self.savedPath)
 
+################################################################################
+# Timeout FUNCTION
+################################################################################
+
+def timeout(func, args=(), kwargs={}, timeout_duration=1, default=None):
+    '''This function will spwan a thread and run the given function using the args, kwargs and 
+    return the given default value if the timeout_duration is exceeded 
+    ''' 
+    import threading
+    class InterruptableThread(threading.Thread):
+        def __init__(self):
+            threading.Thread.__init__(self)
+            self.result = default
+        def run(self):
+            try:
+                self.result = func(*args, **kwargs)
+            except Exception,error:
+                print error
+                self.result = default
+    it = InterruptableThread()
+    it.start()
+    it.join(timeout_duration)
+    return it.result
 
 
 ################################################################################
@@ -1019,6 +1264,117 @@ class digest:
     
 digest = digest().test_all()
 
+#===============================================================================
+# Helper class for timing and RAM flashing of subprocesses.
+#===============================================================================
+class ProcessTimer:
+  def __init__(self,*args,**opts):
+    self.cmd_args = args
+    self.cmd_opts = opts
+    self.execution_state = False
+
+  def execute(self):
+    self.max_vms_memory = 0
+    self.max_rss_memory = 0
+
+    self.t1 = None
+    self.t0 = time.time()
+    self.p = subprocess.Popen(*self.cmd_args,**self.cmd_opts)
+    self.execution_state = True
+
+  def poll(self):
+    if not self.check_execution_state():
+      return False
+
+    self.t1 = time.time()
+    flash = subprocess.Popen("ps -p %i -o rss"%self.p.pid,
+                                              shell=True,stdout=subprocess.PIPE)
+    stdout_list = flash.communicate()[0].split('\n')
+    rss_memory = int(stdout_list[1])
+    # for now we ignore vms
+    vms_memory = 0
+
+    # This is the neat version using psutil
+#    try:
+#      pp = psutil.Process(self.p.pid)
+#
+#      # obtain a list of the subprocess and all its descendants
+#      descendants = list(pp.get_children(recursive=True))
+#      descendants = descendants + [pp]
+#
+#      rss_memory = 0
+#      vms_memory = 0
+#
+#      # calculate and sum up the memory of the subprocess and all its descendants 
+#      for descendant in descendants:
+#        try:
+#          mem_info = descendant.get_memory_info()
+#
+#          rss_memory += mem_info[0]
+#          vms_memory += mem_info[1]
+#        except psutil.error.NoSuchProcess:
+#          # sometimes a subprocess descendant will have terminated between the time
+#          # we obtain a list of descendants, and the time we actually poll this
+#          # descendant's memory usage.
+#          pass
+#
+#    except psutil.error.NoSuchProcess:
+#      return self.check_execution_state()
+
+    self.max_vms_memory = max(self.max_vms_memory,vms_memory)
+    self.max_rss_memory = max(self.max_rss_memory,rss_memory)
+
+    return self.check_execution_state()
+
+  def is_running(self):
+    # Version with psutil
+#    return psutil.pid_exists(self.p.pid) and self.p.poll() == None
+    return self.p.poll() == None
+
+  def check_execution_state(self):
+    if not self.execution_state:
+      return False
+    if self.is_running():
+      return True
+    self.executation_state = False
+    self.t1 = time.time()
+    return False
+
+  def close(self,kill=False):
+
+    if self.p.poll() == None:
+        if kill:
+            self.p.kill()
+        else:
+            self.p.terminate()
+
+    # Again a neater handling with psutil
+#    try:
+#      pp = psutil.Process(self.p.pid)
+#      if kill:
+#        pp.kill()
+#      else:
+#        pp.terminate()
+#    except psutil.error.NoSuchProcess:
+#      pass
 
 
+try:
+    import Foundation
+    import objc
+    NSUserNotification = objc.lookUpClass('NSUserNotification')
+    NSUserNotificationCenter = objc.lookUpClass('NSUserNotificationCenter')
 
+    def apple_notify(subtitle, info_text, userInfo={}):
+        try:
+            notification = NSUserNotification.alloc().init()
+            notification.setTitle_('MadGraph5_aMC@NLO')
+            notification.setSubtitle_(subtitle)
+            notification.setInformativeText_(info_text)
+            notification.setUserInfo_(userInfo)
+            NSUserNotificationCenter.defaultUserNotificationCenter().scheduleNotification_(notification)
+        except:
+            pass
+except:
+    def apple_notify(subtitle, info_text, userInfo={}):
+        return
